+++ thesis_core/src/main/java/thesis/core/experimental/CellBelief.java
@@ -2,14 +2,14 @@ package thesis.core.experimental;
 
 public class CellBelief
 {
-   private float tgtProbs[];
-   private float tgtHdgs[];
+   private double tgtProbs[];
+   private double tgtHdgs[];
    private long timestamps[];
 
    public CellBelief(int numTgtTypes)
    {
-      tgtProbs = new float[numTgtTypes];
-      tgtHdgs = new float[numTgtTypes];
+      tgtProbs = new double[numTgtTypes];
+      tgtHdgs = new double[numTgtTypes];
       timestamps = new long[numTgtTypes];
 
       reset();
@@ -17,7 +17,7 @@ public class CellBelief
 
    public void reset()
    {
-      float equalProb = 1.f / tgtProbs.length;
+      double equalProb = 1d / tgtProbs.length;
       for(int i=0; i<tgtProbs.length; ++i)
       {
          tgtProbs[i] = equalProb;//Assume equal probability of all target types
@@ -26,12 +26,12 @@ public class CellBelief
       }
    }
 
-   public float getTargetProb(int tgtType)
+   public double getTargetProb(int tgtType)
    {
       return tgtProbs[tgtType];
    }
 
-   public float getTargetHeading(int tgtType)
+   public double getTargetHeading(int tgtType)
    {
       return tgtHdgs[tgtType];
    }
@@ -41,13 +41,13 @@ public class CellBelief
       return timestamps[tgtType];
    }
 
-   public void updateTargetEstimates(int tgtType, float prob, float hdg, long timestamp)
+   public void updateTargetEstimates(int tgtType, double prob, double hdg, long timestamp)
    {
       tgtProbs[tgtType] = prob;
       tgtHdgs[tgtType] = hdg;
       timestamps[tgtType] = timestamp;
 
-      float total = 0;
+      double total = 0;
       for(int i=0; i<tgtProbs.length; ++i)
       {
          total += tgtProbs[i];
