+++ thesis_core/src/main/java/thesis/core/serialization/world/UAVStartCfg.java
@@ -9,7 +9,7 @@ public class UAVStartCfg
 {
    private int type;
    private WorldCoordinate location;
-   private float orientation;
+   private double orientation;
 
    public UAVStartCfg()
    {
@@ -33,12 +33,12 @@ public class UAVStartCfg
       return location;
    }
 
-   public float getOrientation()
+   public double getOrientation()
    {
       return orientation;
    }
 
-   public void setOrientation(float angle)
+   public void setOrientation(double angle)
    {
       orientation = angle;
    }
